Fix lightbar persistence and reimplement sixaxis motion sensors
Lightbar Changes:
- Add flash storage persistence for DS4 lightbar color/brightness
- Settings now save to NVS when user changes them via START+R2 combo
- Settings load from flash on DS4 reconnect
- Color persists across power cycles, not just Bluetooth reconnects

Sixaxis/Gyro Fixes:
- Fix critical data type overflow bug: int16_t → int32_t for motion sensors
  * Bluepad32 provides calibrated int32_t values (~16384/G for accel, ~1024/deg/s for gyro)
  * Previous int16_t caused truncation and overflow, keeping values stuck or jumping
- Implement proper scaling from Bluepad32 to PS3's 10-bit format
  * Accelerometer: 16384 units/G → 512 units/G (PS3 10-bit)
  * Gyroscope: 1024 units/deg/s → ~2.56 units/deg/s (PS3 10-bit)
- Add byte swapping for PS3's little-endian 10-bit format (0xFF01 = neutral 511)
- Remove broken low-pass filter that was initializing to 0 and keeping values stuck
- Add detailed comments explaining the data format and conversion

Technical Details:
- Bluepad32 DS4 parser provides calibrated values:
  * Accel: ~16384 per G (from DS4's BMI055 IMU, 12-bit accel)
  * Gyro: ~1024 per degree/second (from DS4's BMI055 IMU, 16-bit gyro)
- PS3 expects 10-bit values (0-1023) with neutral at ~512:
  * Values stored as uint16_t with byte swap (0xFF01 for neutral)
  * Accelerometer: 512 = 0G, deviations = acceleration
  * Gyroscope: 512 = no rotation, deviations = angular velocity

+++ Firmware/RP2040/src/Gamepad/Gamepad.h
@@ -106,10 +106,13 @@ public:
         int16_t  joystick_ry;
         uint8_t  analog[10];
         // Motion sensors (sixaxis)
-        int16_t  accel_x;
-        int16_t  accel_y;
-        int16_t  accel_z;
-        int16_t  gyro_z;
+        // Values from Bluepad32 are int32_t:
+        // - Accelerometer: ~16384 units per G (1G gravity ≈ 16384)
+        // - Gyroscope: ~1024 units per degree/second
+        int32_t  accel_x;
+        int32_t  accel_y;
+        int32_t  accel_z;
+        int32_t  gyro_z;
         // Battery level (0-255, 255 = full)
         uint8_t  battery;
